setter for a custom http.Client on Client entity
This is necessary because the mechanisms for setting the timeout
on the underlying http.RoundTripper for the *http.Client is based on
it's discrete type.

Gogole appengine uses a different transport mechanism in urlfetch,
so setting the timeout would causea panic due to an invalis type
assertion. With this change, support is builtin for the standard
*http.Transport and appengine's *urlfetch.Transport