feat(Settings): Add ToMap conversion method
This method is added to help people copy settings from one index to
another because we choose to use a `Map` (`map[string]interface{}`) for
the `SetSettings` and a `Settings struct` for the `GetSettings`.

This new method should only be used on `Settings struct` produced by
GetSettings.

For more informations, please refer to:
https://github.com/algolia/algoliasearch-client-go/issues/108