fix: Set URL parameters for GET requests instead of using the body
A single method of this API client is currently passing parameters with
via a GET call: the `Client.GetLogs` method. As it is a GET, the
parameters should not be passed via the body of the request but via the
URL parameters.

This commit fixes the issue by adding a method to build a request
containing the URL parameters filled with the given body. In addition,
two new functions are also added, simply to refactor the code and make
it cleaner, to build requests with or without empty body.