fix: Prevent invalid batch requests correctly
To prevent making calls with invalid parameters, we use in some places
extra functions to check the parameters. This is the case for the
content of batch operations.

This commit replaces a `break` with a `return`, to effectively early
return from the checking function with an error set. The bug was a
conjunction of the use of a `break` and the error variable shadowing.