fix: Export missing attributes in Settings.ToMap
The following attributes were missing in the result Map returned by the
Settings.ToMap function:
 * `enableRules`
 * `ignorePlurals`
 * `sortFacetValuesBy`

Also, some fields may not have been properly transformed because the
Settings.clean function was not called to clean some fields with sane
defaults.

This commits fixes all the aformentioned issues and adds a test to
ensure that any new field added to the Settings struct will make the
tests fail in not handled in the Settings.ToMap function.