fix: Improve retry strategy by reseting default hosts after some time
We recently discovered that the implementation of the retry strategy for
the client was lacking an important feature: the ability to reset hosts
marked as down after some time. Because of that, an host that was marked
down could never be retried without instantiating a new client.

This commit is a complete rewrite of the retry strategy implementation
to address this issue. For simplicity, we decoupled the retry strategy
implementation from the transport layer, responsible for issuing the
HTTP calls. This new implementation was written in a concurrent-safe
way, meaning that concurrent calls should not result in UB regarding the
state of the retry strategy's hosts.

As the `context` package is now used to be able to effeciently timeout
individual requets, we now have to deprecate support for version Go 1.6
and before (we were previously supporting Go 1.6 and higher) as this
package was only introduced in the standard library of Go 1.7.