feat: Enable debugging output if ALGOLIA_DEBUG is set
To better help while debugging, the client now logs on the standard
output debugging information if the ALGOLIA_DEBUG environment variable
is set to 1 or 2. Using 1 will have the same effect as 2, except that
the output is truncated to a fixed-length to avoid polluting the output
if not needed.

If ALGOLIA_DEBUG is not set, the overhead of the `debug()` statements is
negligeable.