fix: Add missing types for `facetFilters`
Thanks to one customer report, we discovered that the documentation was
wrong regarding the way to set the `facetFilters` parameter. This commit
adds the support for other types (namely `[][]string` and
`[]interface{}`), so that all the following calls are now legit in Go:

```
// category:Book
"facetFilters": "category:Book"

// category:Book AND author:J.K. Rowling
"facetFilters": []string{
	"category:Book",
	"author:J.K. Rowling",
}

// (category:Book OR category:Movie) AND (author:J.K. Rowling OR author:Stephen King)
"facetFilters": [][]string{
	[]string{"category:Book", "category:Movie"},
	[]string{"author:J.K. Rowling", "author:Stephen King"},
}

// (category:Book OR category:Movie) AND author:J.K. Rowling
"facetFilters": []interface{}{
	[]string{"category:Book", "category:Movie"},
	"author:J.K. Rowling",
}
```

Related to https://github.com/algolia/doc/pull/1760.