changed: migrate and add new Insert* methods
The InsertOrReplaceOption function was previsouly living in
algolia/internal/opt/ but has now be migrated to algolia/opt/ to avoid
import cycles.

Also, InsertExtraURLParam and InsertExtraHeader functions were added in
the same place to simplify the code of inserting extra URL parameters or
headers in the option list argument passed down to the Transporter.