perf: improve performances of objectID JSON tag detection
This commit adds a reflection-based implementation to the `getObjectID`
as a first intent to improve performances of objectID detection when the
`objectID` is defined as a JSON tag. Base implementation using
`json.Marshal/`json.Unmarshal` is kept for backward-compatibility and to
ensure that structures for which `json.Marshal` is implemented will
continue to be used for objectID detection.

For a detailed comparison of the different implementations as well as
benchmarks, please refer to our previous commit or to the attached
Github PR: https://github.com/algolia/algoliasearch-client-go/pull/500.