refactor(secured-api-key): remove errs.ErrValidUntilInvalid
The `errs.ErrValidUntilInvalid` error was introduced to signify that the
`validUntil` string parameter of a secured API key could not be parsed
as an integer by the `Client.GetSecuredAPIKeyRemainingValidity()`
method. However, since this string value is always extracted thanks to a
regexp looking for a integer of maximum length 10, the conversion can
never fail, making this error case useless.

This commit hence removes this useless error which get never returned.

[changelog]