feat(transport): expose the default HTTP client via transport.DefaultHTTPClient()
transport.DefaultHTTPClient exposes the default *http.Client used by the
different Client instances of the Algolia API client.

Most users should not need to access this http.Client.

This helper is only useful for users who wish to keep the default behavior
of the Algolia API client (by wrapping it in a user-defined instance of
transport.Requester) but would like to wrap it into a middleware layer or pass
it to an HTTP interceptor.

[changelog]