fix: implement correct type for `Settings.UserData`
So far, the inner value of `opt.UserDataOption` was incorrectly typed as a
`map[string]interface{}` while the engine can actually accepts both JSON
objects or JSON literals.

This commit ensures that any literal can now be accepted as a valid input of
`opt.UserData(...)` (by accepting an `interface{}` instead of a map).

To also be able to correctly decode the inner value of the
`opt.UserDataOption`, the method `opt.UserData.Decode(..)` was also
added to let the users decode the inner value into any Go variable.