fix(search): correctly encode index name in URIs (#574)
We've recently identified that the user-provided index name, when passed
to the URL internally, was not URL-encoded as one would expect. This
could result in the user targetting wrong indices, either when using the Go
client, or other API clients, including frontend integrations.

After multiple internal discussions, we finally decided to introduce
this change, even though it may be breaking for some users, in a minor
version. We consider that the future benefits of fixing this issue for
future new users holds more positive impact than the negative one of
having existing users to have to manually ensure their index names are
correct.

Close #574