refactor(transport): ensure early termination upon context cancellation
Because we were not checking if the error returning by the underlying
requester, we could not determine if the HTTP call failed because of
user-defined context being cancelled because of a timeout being reached or
an exceeding deadline.

This commit introduce an early return for those cases. This way, we do
not perform multiple retries against other hosts when the user-defined
context is already done, because of a context timeout or a context deadline.

Note that we did not use the error wrapping mechanism from Go 1.13 with
`errors.Is()` and `errors.As()` since we are still supporting Go down to
Go 1.11 for now.

Close #596