fix(transport): correctly expose errs.AlgoliaErr when opt.ExposeIntermediateNetworkErrors is passed
**Summary:**

As reported by @sfriquet, the new `opt.ExposeIntermediateNetworkErrors()`
request option was not exposing Algolia engine errors. This was due to
an incorrect assumption that the error returned by the HTTP `.Do()` call
would be non-nil, which isn't in that case (for instance, a 500 does
return a nil error whereas the response itself is a 500 with a response body).

This commit ensures Algolia engine errors are correctly collected when
`opt.ExposeIntermediateNetworkErrors(true)` request option is passed.

**Testing plan:**

A new test has been added: a custom requester, which always return a 500
response without any error, is used to confirm that errors are correctly
reported when `opt.ExposeIntermediateNetworkErrors(true)` is given to
any client function.