chore(circleci): use matrix (#698)
| Q                 | A
| ----------------- | ----------
| Bug fix?          | no
| New feature?      | no  
| BC breaks?        | no     
| Related Issue     | no
| Need Doc update   | no


## Describe your change

This PR uses a matrix for the different go version we want to build. It avoids copy/paste the entire job for each version. The idea behind this change is to eventually split the unit from the integration tests. We will be able to run the unit tests on forked PRs. We have change the required checks from `test-1-XX` to `test-1.XX`.

## What problem is this fixing?

N/A