feat(recommend): implement Recommend-level waitTask (#730)
| Q                 | A
| ----------------- | ----------
| Bug fix?          | no
| New feature? | yes <!-- please update the /CHANGELOG.md file -->
| BC breaks?        | no     
| Need Doc update   | no

## Describe your change

Support waiting tasks with a Recommend-level scope. It is slightly
different than the index-level task because it expects a model name in
the URL. Note that this is historical though, the model name is now
useless and both index-level and recommend-level actually work similarly
behind the hood.

## What problem is this fixing?

Add the abillity to build recommend rules in Metis.