fix(transport): only compress request bodies if compression is enabled (#778)
| Q                 | A
| ----------------- | ----------
| Bug fix?          | yes
| New feature?      | no
| BC breaks?        | no
| Need Doc update   | no

## Describe your change

Only compress the body of an HTTP request if compression is enabled.

Without this change, the content of the body can get compressed even
though no `Content-Encoding` header is added.
What's more, this also causes empty bodies to result in a body in the
HTTP request. That's an issue for `DELETE` and `GET` calls for example,
where some server implementations reject any request with a non-null
body.

The changed is covered by a new non-regression suite. Without this fix,
the `DELETE gzip compression server-side body` test case would fail.